Refactor common utilities for creating models
